Add more logging around the persistent connection
Summary:
I occasionally observe an issue where we have a healthy Flow server and an
apparently healthy IDE process which has received a subscribe request.
Unfortunately, somewhere between the IDE process and the Flow server, there is a
bug where the message gets dropped. Adding logging to help pinpoint it.

Reviewed By: gabelevi

Differential Revision: D4813104

fbshipit-source-id: b8b11d3be41b31b612cc6543453f05508b42a320