[hack] implement textDocument/rangeFormatting
Summary:
This implements range-formatting.

(I've put in parsing for document-formatting as well since the two share lots of parsing, but it will require heavier work, which I'm leaving for a future diff.)

This current diff involved fixing an off-by-one bug in ServerFormat.go_ide. That was only ever invoked in one place (clientIde.ml) and had never been hooked up to Nuclide.

Reviewed By: arxanas

Differential Revision: D4805304

fbshipit-source-id: ec24b7014ce2bcbf6f204dc2d3f7f6c403e09043