[hack] implement textDocument/formatting
Summary:
The bulk of this diff is changing ServerFormat.go_ide so that:

(1) it accepts a 'None' range, indicating that you wish to format the entire document. (Previously when we formatted the entire document we got Nuclide to tell it the range of the whole document. But now with LSP, the LSP textDocument/formatting, that information isn't communicated to hh_client. And hh_client can't do this itself since it lacks knowledge of document contents. That's why we have to pass 'None' to server to leave the server to figure it out.)

(2) it reports what range it ended up formatting. For now, this is essential in the whole-document case, so that LSP textDocument/formatting can report back a TextEdit with the range of the document that has been formatted. In future, I hope that hackfmt will intelligently expand its range, e.g. to encompass whole statements, and report back what range it formatted.

For future, we'll be able to rely wholly upon hackfmt to do the formatting and report back a TextEdit. But up until the switchover happens, and we're still using hh_format, then we can't do that yet.

This change had a few impacts on clientIde.ml, in the way it deals with formatting. But we're not using clientIde formatting at the moment, and we're not going to in the future, so I did just the bare minimum to make it compile cleanly.

Reviewed By: dabek, arxanas

Differential Revision: D4808804

fbshipit-source-id: 9266c5fe4fb5f4b0a1977901de4a0f72055992e5