`export type * from ...`
Summary:
I never implemented this because nobody ever asked for it, but I assumed it would pop up eventually.

This is the `export type` version of `export * from "blah";`. It takes all the type exports of the source module and forwards them as exports of the current module. It omits any exports that would not be compatible with `import type`, which is what sets it apart from `export * from` (which forwards *all* exports -- both type and value).

leeb asked for this over in https://www.facebook.com/groups/flowtype/permalink/1394486823933262/

Fixes #3152

Reviewed By: samwgoldman

Differential Revision: D4863922

fbshipit-source-id: 6205878940159a457d94711e2a30d8978e8f1255