Refactor Daemon module to allow passing in stdin
Summary:
We want to read from stdin in the background. Threads didn't work when I tried them (segfaulted in the garbage collector), so I'm sticking with processes, which is what's used everywhere else. In order to read from stdin in a background process, we need to have the child process inherit stdin.

The easiest way would be to call `fork` and pass in the stdin to inherit. However this isn't portable to Windows. Instead, we have the `Daemon.spawn` which uses the portable primitive `Unix.create_process`. This requires a lot of machinery to pass data to the child process. Rather than reimplementing the `Daemon.spawn` system, it's probably cleaner to just support inheriting stdin.

Reviewed By: dabek

Differential Revision: D4644356

fbshipit-source-id: d1db6dbc07afea32e3c49946e1e18fab39a5421e