[hack] ClientBuild has force-dormant-start option
Summary:
Waiting for repo settling will be moved from the server init into the monitor.
So starting a server will be put on hold if the monitor detects an
update/rebase is in-flight during startup.

This poses a problem if Hack is started up via Hack Build from Mercurial's
mergedriver. To get past this blockage, we add a force-dormant-start option
to Hack Build, which the merge driver needs to start using (once this is
deployed)

Reviewed By: dlreeves

Differential Revision: D4851204

fbshipit-source-id: 3374959941fc24b5c3ac32aa9120b18ef458557a