Fix incremental bug when lib files are added or removed
Summary:
This diff fixes a couple errors that prevented recheck from detecting changes
when lib files are updated.

First, adding a libdef could override a builtin definition or definition from
an "earlier" libdef. In `process_updates`, when we see changes to lib files, we
kill the server. However, we only tried to find updates in lib files we have
already seen, so new lib files did not cause the server to die.

This diff looks at both existing and new lib files. We need both because lib
files could have been added or removed.

Second, reparsing lib files was not handled properly. Lib files should always
be parsed in types mode and strict mode, as we do in `init_lib_files`. However,
`reparse_with_defaults` was not passing those flags correctly.

This diff updates the signature of `reparse_with_defaults` so we can pass in
the same overrides we use during init for lib files.

Reviewed By: mroch

Differential Revision: D4881107

fbshipit-source-id: 850271522c8105aa9dc1692e4fc85e7f89222a74