Fix parser locations for unary operators
Fixes #3744

Previously, when determining the end location of a unary operator, the parser
would just the the end location of its operand, but this is wrong if the operand
is surrounded in parens. To fix, we can use `with_loc` to get the full location
including parens, like is done for binary operators.