[PR] Fix parser locations for unary operators
Summary:
Fixes #3744

Previously, when determining the end location of a unary operator, the parser
would just be the the end location of its operand, but this is wrong if the operand
is surrounded in parens. To fix, we can use `with_loc` to get the full location
including parens, like is done for binary operators.
Closes https://github.com/facebook/flow/pull/3745

Reviewed By: mroch

Differential Revision: D4899751

Pulled By: nmote

fbshipit-source-id: 90d064a259bddc95fcb9b3501ef125c068b3cba5