[PR] Fix the travis deploy
Summary:
`make build-flow-with-ocp` would build Flow with ocp-build, but wouldn't do the `objcopy` step which would copy the flowlibs into the binary. Since it was performed last, it was overriding the `bin/flow` binary that we had previously tested. Then, on tagged commits, we would publish the bad binary.

This makes two main changes:

1. Use `make all-ocp`, which does the full ocp-build build.
2. Move the ocp-build build before we test, so that whatever is in `bin/flow` is properly tested.
Closes https://github.com/facebook/flow/pull/3756

Reviewed By: mroch

Differential Revision: D4906157

Pulled By: gabelevi

fbshipit-source-id: 23c11080199652382bd21479ac4b76fdeb4656d6