let dfind process exit on its own
Summary:
copied this comment from Hack, and in my tests it appears to be true. I
added a `die ()` to the end of `print_status`. I ran `flow start tests/logical`
and saw that the dfind process was running (found the pid in the pid log). Then
ran `flow status tests/logical` which hit the `die ()`, and saw that the
workers, dfind server, and master all exited.

I want to do this because it's the only thing outside of the recheck loop that
interacts with `dfind`.

Reviewed By: nmote

Differential Revision: D4903392

fbshipit-source-id: 7f3e056c183c0f2aa03ad254184d0ad2ec0197c7