remove redundant lock check
Summary:
we grab the main lock in `ServerInit.run`, and then call `MainInit.go`
which tries to grab the lock again, which is a no-op.

I suspect that we used to grab the lock only in `MainInit.go`, but changed to
grab it earlier and didn't remove the later check.

Reviewed By: gabelevi

Differential Revision: D4900113

fbshipit-source-id: ac647ac917f90597df61a0d914badb433053a033