Send the end-recheck message before sending the final set of errors
Summary:
The plan is to send the complete set of errors at the end of a recheck, to
mitigate the negative impact of bugs in the code that will send errors down as
they are computed. While such a bug would impact the intermediate results, at
least it will have no effect once the recheck is complete.

Nuclide will treat any update appearing within the recheck bookends as an
incremental update, and will simply add it to previous results rather than
replacing previous results with it. By moving the end_recheck message before the
final error update, Nuclide will treat this final update as the source of truth
and will replace all previous errors with those contained in this update.

Perhaps it will make sense to make this distinction (incremental vs complete
error set) more explicit in the future. Certainly we will have to figure out how
this scheme can be made compatible with the LSP. Fortunately the "very-unstable"
protocol name gives us quite a bit of leeway to change things :)

Reviewed By: avikchaudhuri

Differential Revision: D4918825

fbshipit-source-id: 49506a0a2f233eed1928047ffd22fd67e72a7438