Fix prop name in "Optional object properties" example
The third example given in the "Optional object properties" section on this page is supposed to produce an error, but doesn't. I verified this using the "Try Flow" page, as well as a new repo with a minimal Flow setup installed.

It looks like the prop name used in the function signature is unintentionally different from the one used in the function calls. Changing `optionalProp` to `foo` causes the third example to produce an error, as expected.