calculate requires for an ast
Summary:
Walk an AST to calculate its requires.

Previously, this step was part of infer_ast. However, calculating requires is
necessary to construct the dependency graph of a project, which in turn can be
used to drive lazy type checking (in particular, not performing infer_ast on
some files). The dependency graph could also be saved and used to "boot up" a
type checking session.

Note that the requires calculated by this diff does not exactly match the
requires calculated by infer_ast in all cases. Consider all such differences to
be bug fixes. In partcular, we were not recording requires after a
require.XXX(...)  call, or when esproposal_export_star_as was false.

(Note: this ignores all push blocking failures!)

Reviewed By: samwgoldman

Differential Revision: D4816349

fbshipit-source-id: 2a666c8d1d80290087ce26ed2ff1f3205215511f