node modules opt
Summary:
Major optimization of how requires are resolved.

Previously, each require in a file would be resolved by walking up the file
system, testing whether parent directories contained node_modules. (Actually,
the process is even more inefficient, but this is a good approximation.) This
effectively causes the file system to be stat-ed over and over again, multiple
times for each file, for all files, in parallel.

This diff instead remembers node_modules "containers" (parent directories) when
walking the file system before parsing, and uses that information to jump or
bail early. This doesn't get rid of all the inefficiency in the process, but
does eliminate the biggest by far.

One test suffers, where in addition to reporting a required-module-not-found we
additionally diagnosed it being outside the root. Will figure it out later.

Reviewed By: samwgoldman

Differential Revision: D4854857

fbshipit-source-id: 72da811521814de9189cc189542b164c3b7698d3