Fix default constructor bug.
Summary:
A class with no constructor gets a default constructor. However, if it
extends another class, we omit the default constructor. This is right and good.

`mk_interface` was missing this logic, which meant a `declare class` without a
constructor would always get a default constructor and ignore any inherited
constructor. This is bad, and fixed by this diff.

A slightly orthogonal question is whether we should give interfaces default
constructors at all. We currently do, but that actually seems a little weird.
Maybe in a different diff I'll try and change that behavior.

Reviewed By: samwgoldman

Differential Revision: D4935970

fbshipit-source-id: 750ebaba36f11df9ef1289875b7b9c2413ca7160