Use RIdentifier instead of RCustom in convert_qualification
Summary:
Future diffs will change the reason of all env lookups to be `RIdentifier`
instead of the hodge-podge of RCustom reasons we have now.

I singled this one out because it has an unexpected side effect in the tests.

This should be a simple change, but causes a frustrating change in error output.
The perturbation is caused by a change in sort order of reasons. This causes
constraints to fire in a different order, which causes Flow to error during
speculation in the call to $await.

Sort order of reasons should not cause a change in error output, but just
because it's bad here doesn't mean it's not better elsewhere. We should fix the
underlying issue, but I don't think we need to block this.

Reviewed By: avikchaudhuri

Differential Revision: D4920942

fbshipit-source-id: 70c93b6fb3682257861a9befc907c231c2c99afe