Use Loc.t instead of Reason.t for cache_global
Summary:
This is the Big Change(tm) in this stack. We thread a reason through a bunch of
Env functions in case they fall through to global lookup. While a variety of
reasons did reach this position before, it's more convenient and more consistent
to use RIdentifier for all of them instead.

One unfortunate regression is for not-found jsx pragma replacements. Since we
search for the replacement identifier name, the error message says "Foo" can't
be found and points to <Bar />.

I really want to get rid of the nonsense RCustom reasons we create for no good
reason (hah), so I'm inclined to accept this regression.

Reviewed By: gabelevi

Differential Revision: D4920992

fbshipit-source-id: d5af5431214f581e4645d1d61f22e5d50359d90a