Add optional override to default RIdentifier desc on not found env refs
Summary:
I'm convinced that it's useful to provide a specific description in exceptional
scenarios (like JSX desugaring).

Requiring the desc argument led to lost of useless RCustom descriptions, though,
so I want to make the argument optional.

This diff restores the description for not-found JSX elements in the presence of
a `jsx` pragma rewrite rule.

Reviewed By: gabelevi

Differential Revision: D4948008

fbshipit-source-id: a4f21fb2fd5a21c68eb51c7f4046f07dbaf5631e