[PR] Print out usage guidelines when non-numeric line & column are given to type-at-pos
Summary:
Hello!
I noticed that the error feedback when using `type-at-pos` incorrectly could be a little more consistent.

In the current build, passing a filename and line but no column results in the following output:
```bash
$ ./bin/flow type-at-pos some/file 2
Unhandled exception: Failure("int_of_string")
```
This is because with two arguments `type-at-pos` expects to receive line and column, and for the filename to be given via stdin. Most filenames can't be parsed as ints, hence the `Failure`.

The same error occurs if three args are passed, but the second two are non-numeric, e.g.
```bash
$ ./bin/flow type-at-pos some/file 2 a
Unhandled exception: Failure("int_of_string")
```

This change just makes `type-at-pos` bail out if the line & column are unparseable with the same behavior used for the zero arguments case.

Please let me know if there's anything I can do to make this better!
Closes https://github.com/facebook/flow/pull/3834

Differential Revision: D4976472

Pulled By: gabelevi

fbshipit-source-id: 546b5edc6f737266b33fe910edf77deea98df83a