get-def rewrite
Summary:
Flow's get-def behaves differently than TypeScript's (observed through vscode,
e.g.) for locals. A rvalue points to the lvalue of the last assignment, the
latter points to the value being assigned. Basically we can trace the flow path
for a local variable.

Unfortunately, this doesn't match expected behavior. Following TypeScript, this
diff now makes it so that for locals, rvalues as well as lvalues of assignments
point to the declaration site for that variable.

In addition, we retain a special case of the tracing behavior: when pointing at
a variable declaration site, instead of not doing anything, we jump to the
initializer's definition site. This seems generally useful, but also in
particular it allows us to model the following:

const foo = require('./foo')

Pointing at `foo` on the left of the `=` takes us to the exports of whichever
file `'./foo'` resolves to.

This corresponds to similar behavior in TypeScript for the following (CommonJS/AMD):

import foo = require('./foo')

(even though for other variable declaration sites, TypeScript does nothing).

The remaining case is ES6 modules:

import { foo } from './foo'
foo; // use of `foo`

where the use of `foo` should jump directly to the exports of whichever file
`'./foo'` resolves to. This behavior is already in TypeScript (but not for the
CommonJS/AMD case). Will follow up with an implementation of it in Flow.

Reviewed By: samwgoldman

Differential Revision: D4927659

fbshipit-source-id: deff7c8419f4753e95880e66c44a2135682913fd