definition locations
Summary:
This diff introduces another field in the reason record, marking the optional
"def_loc" of a type.

The idea is that a type, when created, might have a location that gets lost by
repositioning, so the "loc" field of the type's reason might not point to the
location of its definition (it points to its use).

As an optimization, we initialize "def_loc" to None (so that "loc" holds the
location when the type is created), and only save "loc" to "def_loc" when "loc"
is changed (on repositioning).

Reviewed By: samwgoldman

Differential Revision: D4927601

fbshipit-source-id: ba041e727f7b12560936b6728141c254e5a76cf6