[PR] Changes dynamic import syntax to work without assignments
Summary:
Looks to me that the parsing only looks at the `import` only supports static imports and not dynamic imports, as it's matching the token and parsing it as a import statement, even if the next token is a parenthesis. But when it is dynamic it should be treated as an expression.

This matches up with how babylon6 ast parser does it: http://astexplorer.net/#/gist/4dcc5f44e952f664c4e9f4c9098b864c/3db751c05b8b6ce7dd0f813f913c0c79efd7b120

This breaks with calling `import()` without `.then()`, but as promises are eager does really observing the result make a difference? I couldn't find any information on whether or not this is not supported. I know `import` isn't strictly a function, but a keyword with parenthesis, so maybe this changes the semantics. If this is the case we need to see if there is an identifier property as a port of the expression.

Looks like deecewan added the initial testcase. Maybe you have some inside into if `import('dsa')` is valid or not?
Closes https://github.com/facebook/flow/pull/3844

Differential Revision: D4979054

Pulled By: gabelevi

fbshipit-source-id: c3f38c839315a2bc6487bb00af8f0a23a99b3976