[hack] Make live global errors a bit less live, but more performant
Summary:
The way it works now: the moment you run global analysis and get thousands of errors, we send only up to 50 of them to editor (and from now on prioritize recomputing them), keeping the rest hidden. If you fix one of them, we'll "top up" with the 51th, etc. until you fix them all.

This keeps the live errors promise ("if something is displayed in editor, we try to keep it live and up to date"), while (in theory) also providing a way to handle "thousands of errors" use case without running expensive global check over and over.

Unfortunately, since D4129071, due to having to fix a correctness issue, maintaining the "hidden rest" up to date incurs a significant (smaller than global recheck, but proportional to number of errors) cost. This leads to complaints from framework developers.

In this diff, I get rid of the "drip errors one by one" part - we just forget about most of the errors in unopened files, and you need to re-run global analysis to top up again after you fix the ones you got. I see a way forward to automate this with some heuristics, but for now it seems like a strict improvement for framework developers, with minimal disadvantage for the rest.

I also lower the number of reported unopened files with errors from 50 to 10 (I don't think that there is any usability difference...)

- changes in ServerTypeCheck.ml are mainly a revert of D4129071.
- a bunch of test files change, because the new approach also implements the filtering solution from D4775042 better: previously we would de-duplicate errors only in non-open files, now we de-duplicate all errors. I checked that all of the changes are deletions of diagnostics that were re-sent unnecessarily
- the scenario that we are changing, error throttling, is in test_error_throttling.ml: I change it by adding another STATUS (i.e. global recheck) to the test

Reviewed By: jamesjwu

Differential Revision: D4974336

fbshipit-source-id: 93ff3a2026b674ede03a3ffa6f991d250c5aa809