Show up to 5 lines of context in error pretty printing
Summary:
Previously, if an error location pointed to more than a single line of
code, we'd just point to the beginning of the error. This diff changes that
behavior to show up to 5 lines of context. If the error spans more than 5
lines, then we show the first 3 lines, a `...`, and then the last line

In order to indicate where on the first line the error starts, this diff also
adds an overline of `vvvvvv`.

Reviewed By: samwgoldman

Differential Revision: D4947465

fbshipit-source-id: 681d2f73e231b735a8278256ba25e086f677ccbf