Ensure React.createClass tvars are 0->1
Summary:
We use 4 tvars when analyzing React.createClass, representing `this`, the static
`this`, and the `state` and `default` type params. We create these tvars first
because they have a cyclic definition.

These types are determined by a single lower bound, which is calculated at the
end of the analysis.

If any of these types appear in the opposite position of a speculative
match--i.e., as the lower bound of a union or the upper bound of an
intersection--we should be able to wait for their lower bound and continue.

Before this change, the speculative matching machinery would not wait for the
lower bound. Instead, it would turn the tvars into any. By using "instantiable"
reason descriptions, Flow knows that we can wait for a single inflow.

Reviewed By: avikchaudhuri

Differential Revision: D4995300

fbshipit-source-id: 288f8200f9e3b511f5fb2552a6596b0cb52dfc79