Fix inconsistency prop vs. p
Text referred to property 'p' but code had 'prop'.  This change fixes that inconsistency.

Besides making them consistent, I propose 'who' instead of either 'prop' or 'p'.
Also, I find it confusing to have variables named 'employee' and 'person' that are not of type 'Employee' and 'Person', but rather contain a property 'p', or 'prop' that is of those types. I think calling the property 'who' alleviates the problem somewhat.