Make autocomplete function return results rather than marshalling them directly
Summary:
In order to do autoocomplete over the persistent connection, we need to wrap the
result in a variant. The client cannot simply assume that it will be an
autocomplete result.

If this looks good I'll follow up with diffs to convert every case in respond to
use this pattern.

Reviewed By: avikchaudhuri

Differential Revision: D5005662

fbshipit-source-id: 49e590341dec0f7d815f73a339df8afedfde1b7c