Store logging_context in persistent connection record
Summary:
This is required in order to call several functions in server.ml, including autocomplete, get_def, etc. which we will want to be able to do over the persistent connection.

I'm not 100% sure this is the right way to go. I'm not sure how this information is ultimately being consumed. Would it be confusing to have a bunch of analytics that include the `argv` and start time, etc. of the long-lived ide command? Maybe it makes more sense just to refactor this argument to be an option, where it is currently required?

Reviewed By: gabelevi

Differential Revision: D5012177

fbshipit-source-id: e9f1d37f2474166bd0a499e1d8f739bd4693a772