[PR] Simplifying & Clarifying Module Docs
Summary:
The examples shown in "Importing and exporting types" are inconsistent with the examples shown in the other sections. The exports from "exports.js" do not correspond with the imports from "imports.js", thus confusing any reader reasonably assuming correlation. In addition, the examples are duplicative of the first section.

This diff merely consolidates the first two sections and maintains a coupling of "exports.js" to "imports.js" to help the reader understand the relationship.
Closes https://github.com/facebook/flow/pull/3865

Differential Revision: D5020819

Pulled By: avikchaudhuri

fbshipit-source-id: 78df122083cbd66acd89b86bd7a2f87f86273e99