Allow URL instances in RequestInfo
A URL instance is a valid argument to `fetch(url)` and `new Request(url)`.
It's not mentioned explicitly in the spec:

    typedef (Request or USVString) RequestInfo;

But, technically, any non-Request object that implements `toString()` is
a valid argument, since it will be converted into string:
https://github.com/whatwg/fetch/issues/452#issuecomment-272439900

However, I didn't want to allow `any` as a type for RequestInfo, since
that basically eliminates the benefit of typechecking the first argument.
I'm not sure what the best approach to this would be, so in the meantime
I have just added "URL" to the RequestInfo union.