master context cleanup
Summary:
cleans up how the master context is created. now it's created by
`Init_js.init` where it is first used, instead of being pre-created in
`Server.preinit`.

It is created on the master and written into ContextHeap for the workers to read
during merge. This was already the case, but it was also passed to the forks.

I also removed some callers of `get_master_cx` to be able to inline that into
`Init_js.init`. They were clearing (in `recheck`) and storing (in `typecheck`)
errors in the master cx, but AFAICT, errors are never stored in the master cx
itself so this was all a no-op.

Reviewed By: avikchaudhuri

Differential Revision: D4896016

fbshipit-source-id: 69aebac5c2759615f2de4e8b8e004763ed5add6f