def type refactor
Summary:
Move a number of cases out of Type.t into a new type Type.def_t, which excludes the reason, and which is wrapped by DefT, that houses the excluded reason.

How did I decide which cases to move out? The cases correspond to surface types the user will care about. Internal types are not moved. Some remaining cases might be arguably candidates for moving in the future, but they stick out as having a different flavor in my head right now.

Why this change?

This should make reason-related (and other general, metadata-related) manipulations easier to separate from type-checking mechanics.
For example, extracting and modifying reasons becomes less intrusive. Experimenting with reasonless compare becomes easier.
We can also put things like def_loc (and in the future, trust bits) exclusively on reasons in DefT.

Finally, we can do a pass over partial functions that are only defined on def_t, and make them total.

(This diff doesn't do any of the above "exploits" to keep it a pure refactor.)

Reviewed By: samwgoldman

Differential Revision: D4980282

fbshipit-source-id: 3dbfa8fae64bf0c6867f35484c50fdd0e4539eb7