check reverse dependencies
Summary:
Previously, given a file to focus-check, we would only crawl dependencies
(upstream), not reverse dependencies (downstream). Now we do.

In principle, this completes `focus-check`. We can expand the command to take
multiple files, at which point we should be able to run it on diffs.

The next part is to make this command "live," i.e., build `focus-status`.

(Note: this ignores all push blocking failures!)

Reviewed By: samwgoldman

Differential Revision: D4873812

fbshipit-source-id: 16130ccc0ff5a963c4b3ce7275f5a26a846a4103