split hardcoded parser tests into individual files
Summary:
splits `hardcoded_tests.js` into individual files for each test case.

it actually uses the same test runner as the esprima tests. each test is a
directory, and each case is a `.js` file for the source and a `.tree.json` file
for the expected AST.

this version requires `.tree.json` to be the full AST, whereas previously
`hardcoded_tests.js` only checked some fields. this is more safe, and I think
actually easier to deal with. I added a `-r` option to re-record the
`.tree.json` files, which is much easier than hand-writing the subset of fields
to check.

Reviewed By: gabelevi

Differential Revision: D4982868

fbshipit-source-id: 78764f5728fad2335de7cc5db70daad270bf4034