Added additional OGP tags and renamed image
I’ve renamed the OGP image to have a “-1” so if it is ever updated it’s
easy to replace. As I understand it, the Facebook scraper does not
refetch resources with the same name and recommends changing the name
when changing a resource.
I also added a site name tag, so that each page is tagged as being from
the Flow site.
I added a title and description tag that use the existing meta
information to also provide this to the Open Graph scraper.