[hack] support "tentative" attempt to connect
Summary:
Goal is that (1) if Nuclide's connection to hh_client goes down due to a crash then it should be automatically restarted, but (2) if it goes down due to a separate Nuclide window connecting to the same project, then it shouldn't be automatically restarted.

This had previously been solved by jakebailey in D4537938 HackProcess.js line 304 by having Nuclide observe the exit code of hh_client and, if it was HACK_IDE_NEW_CLIENT_CONNECTED_EXIT_CODE, then declining to auto-restart.

But that approach doesn't fit in LSP, which doesn't pay attention to error code.

Instead, here's a VSCode-friendly LSP-friendly mechanism to accomplish the same thing: https://github.com/Microsoft/language-server-protocol/issues/227

1. When an LSP proxy gets a first Initialize request over its connection, it should first check whether any other LSP proxies are connected to the same singleton backend server. If they aren't, fine, it can proceed. If there are, then it should fail the Initialize request, and give back a ResponseError saying "Sorry another client is connected at the moment" with the retry flag set to true.

2. When an LSP proxy gets a later Initialize request over its connection, then we know it must have come in response to the user explicitly clicking the Retry button (even though LSP spec doesn't explicitly convey this fact, we can still deduce it). So at this time it should kick off any other proxies connected to the same singleton backend server, causing them to terminate, and it should proceed with initialization as normal.

Reviewed By: arxanas

Differential Revision: D5015027

fbshipit-source-id: 1f0acdc01ce1dc38cbee5892418806b76a0296a6