Add children and seal react props literal
Summary:
When analyzing a JSX expression, like `<C p={0} />`, we never sealed the object
type representing the props object (`{p: number}`). This is because we never
included `children`.

While typed children are still WIP, this diff ensures that the children property
is at least part of the object type when there are childen, abeit with an `any`
type.

Unfortunately, this is not sufficient to allow React component's Props types to
be exact. Because Props is defined as the solution to an equation with
existential types: `Config ~> DiffT (Props, DefaultProps)` where `Props` is `*`.

It's an implementation detail, but because `DiffT` is implemented in terms of
`ObjAssign*T`, the `Props` object needs to be open in order to accept the
ensuring `SetPropT`s.

Reviewed By: avikchaudhuri

Differential Revision: D5074775

fbshipit-source-id: baab51fc508aa7fa509e9d98b4dc972576ea8faa