[PR] [Parser] Don't rollback on colons without parens
Summary:
The js_of_ocaml compiled flow_parser.js contains a line with 10k characters, which is mainly just a ton of nested ternary expressions. That means there are a ton of expressions followed by colons.

When parsing an assignment expression, the parser generally works like this:

1. Try to parse an assignment expression which is not an arrow function expression. If there are errors or if the expression is followed by `:` or `=>`, then rollback
2. If we rolled back, then try to parse an arrow function expression. If there are errors, rollback
3. If we rolled back, then parse an assignment expression that is definitely not an arrow function.

The "rollback if the expression is followed by an `:`" rule was causing lots of rolling back. We can work around this by only rolling back if the expression ended with a paren AND is followed by a colon. This is because return type annotations are not allowed on arrow functions without parens.

  var f = (x): number => 123; // allowed
  var g = x: number => 123; // disallowed

Closes https://github.com/facebook/flow/pull/3963

Reviewed By: samwgoldman

Differential Revision: D5079421

Pulled By: gabelevi

fbshipit-source-id: 9af026be68495d7c25aa83694ef3a1ce1d0e3110