Object library calls
Summary:
Before this diff, certain operations simply did not work on ObjProtoT. This is
mostly OK, because that type most appears in internal constraints which don't
involve GetPropT, etc.

However, it's absolutely possible to create an (ObjProtoT _, GetPropT _)
constraint:

```
({}.__proto__.toString);
```

Before this diff, the above code was an error: "Property `toString` cannot be
accessed on object type."

After this diff, the expression evaluates to a function type.

Reviewed By: gabelevi

Differential Revision: D4907791

fbshipit-source-id: 4ad23fa4d86fb5db0a5f16ea7e6ded53c96f0ad3