remove cli error flags from Options
Summary:
these flags are only used by the CLI error printer, which is only used
by the check and status commands. since the flags are parsed by the check and
status commands, they can be passed directly to the printer instead of being
stored in `Options`.

(Note: this ignores all push blocking failures!)

Reviewed By: gabelevi

Differential Revision: D5069786

fbshipit-source-id: 76fa71cb708c8b85b6b4043d0ad84a3a7c17cd40