remove loc from coverage, type-at-pos, dump-types errors
Summary:
when `coverage`, `type-at-pos` or `dump-types` gets an unhandled
exception, we serialize it to the client so it doesn't get swallowed. but we
also include the `loc` that was passed by the client, which is pretty useless.

the motivation here is just to simplify the code, and remove some callers of
`Errors.deprecated_json_of_loc`.

since this is an exception case that never happens (modulo Flow bugs), I'm not
worried about the minor backwards-compatibility change since it's not really
part of the public API.

Reviewed By: nmote

Differential Revision: D5084422

fbshipit-source-id: d410503a5cb3212bcfab18d0b16bcf9908794b33