move suggest command file diffing from server to client
Summary:
the suggest command works by generating a list of annotations to
insert, then using that to generate a diff as output.

a diff is a nice output for a CLI, but not the best for JSON or other clients.
so, move the diff generation from the server, to the CLI client so that the
server response is display-agnostic.

Reviewed By: nmote

Differential Revision: D5087555

fbshipit-source-id: 3acc04dc80c6a44e5c1a35c350e2130554a16056