[cleanup] rename ServerProt send/receive functions
Summary:
`ServerProt` had 4 functions for client/server message passing, for
a client to send a message, a client to receive a response, a server to receive
a message, and a server to send a response.

I found these function names confusingly similar. So, I moved the 2
server-related functions into `Server`, and the 2 client functions into
`CommandUtils`, and renamed them to be clearer.

Reviewed By: gabelevi

Differential Revision: D5091467

fbshipit-source-id: 057b1a6ce7d4569245c7c9ab82f5e81ee0b7cc63