remove SERVER_DYING message
Summary:
the `SERVER_DYING` variant of `ServerProt.response` was only used by
the `stop` command. other commands had to handle `SERVER_DYING`, even though it
was impossible for them to receive it since we only ever send it in reply to a
`stop` client.

so, removed `SERVER_DYING` and replaced it with `stop_response`, so other
commands don't think they have to handle it.

Reviewed By: avikchaudhuri

Differential Revision: D5101399

fbshipit-source-id: ae3876ff7e79f99b15ba3439b9f010f8e4128900