Model __proto__ read/write as getter/setter on Object.prototype
Summary:
Flow's current support for `__proto__` lookups is not quite right. We intercept
property loads, but the spec states that `__proto__` is actually a getter/setter
on Object.prototype.

```
var o = { ["__proto__"]: 0 };
console.log(o.__proto__);
```

The above code will print `0` and `o`'s prototype is still `Object.prototype`.

Note that the rule for `(DefT (_, InstanceT _), GetProtoT _)` is wrong. It was
wrong before, and it continues to be wrong in the same way. I want to fix this,
but it's hard to do, so I'm punting on it.

Reviewed By: avikchaudhuri

Differential Revision: D5118973

fbshipit-source-id: 838e771727c6ea09df547eea873d0eaaecd8742c