consistently ignore SIGPIPE
Summary:
we should ignore SIGPIPE signals so that an EPIPE exception is thrown
instead, since we can catch and log EPIPE if we want to. we are already ignoring
SIGPIPE in the client's `main()` and in the server's `run_internal` (which is
basically its `main`), so moving it to one place that calls either of the others
should be equivalent.

Reviewed By: gabelevi

Differential Revision: D5124562

fbshipit-source-id: 8e41775b8293b8a21a4308678dc19dc82c79ec39