Limit of one outstanding IDE request
Summary:
If we are sending a lot of data through the Flow client/Flow server sockets, we
can eventually fill up the buffers. If both the buffers fill up at the same
time, we can end up in a situation where both the Flow client and the Flow
server are blocked on writes, and we deadlock.

This diff makes it so that the Flow client will queue up incoming requests and
only send one to the server after the server has responded to the last request.
If we were dealing only with requests, this would solve the problem entirely,
since if the server is in the middle of writing a response, there is no way for
the client to also be doing a write, because the client would be waiting to
receive that response before it issued the next request.

However, we also have push notifications coming down from the server. Therefore
it is possible for the client to write a request at the same time that the
server writes a notification. So, in theory we could still have this deadlock
problem, though at the very least it has been significantly mitigated.

Reviewed By: avikchaudhuri

Differential Revision: D5127088

fbshipit-source-id: 13f56721f21a959f25f1701c20c7ad8a7a30cba5