quick-start implementation
Summary:
Implementation of flow quick-start.

This is a variant of flow start which parses all files and calculates
dependencies between them, but then does nothing until something changes.

When something does change, it does what it would do on a recheck, and a bit
more. The additional work arises from the fact that rechecking files cannot
assume that files that they depend on have already been checked. So dependencies
need to be "rechecked" as well.

Reviewed By: gabelevi

Differential Revision: D5067264

fbshipit-source-id: e39572f2e0d79295e7b7439323930e737482ea75