Clarify primitive promotion
Summary:
These rules are designed to implement the ToObject rule in the standard[0],
which promotes primitives to their object wrappers. Giving the list of use ops
for which we should convert with ToObject makes the intent a bit clearer, and
separates these rules from the similar seeming rules for Array, Function, and
Object.

0. http://www.ecma-international.org/ecma-262/6.0/#sec-toobject

Reviewed By: gabelevi

Differential Revision: D5148217

fbshipit-source-id: 0f4b8fc304e344645de4fa3cd7b91b86ff3a85cd