Swap 'soundness' and 'completeness'; tweak grammar
The definitions of soundness and completeness were the wrong way round, see:
http://logan.tw/posts/2014/11/12/soundness-and-completeness-of-the-type-system/
https://stackoverflow.com/questions/21437015/soundness-and-completeness-of-systems

and other resources.

I took the liberty of tweaking a couple of bits of grammar/wording while I was here too.

On a separate note, the sentence "Unit/Integration testing can be an extreme form of this approach." is thrown in without any elaboration. Although I can see how you could make this argument / draw this analogy, I think it could do with expansion/justification. Happy to write a separate change for this if people agree.

Thanks. :)