split context metadata into local and global options
Summary:
many of the fields in `Context.metadata` are actually global options,
the same for all files and across calls to the server.

instead of writing them into `Context.t` which is read and written to sharedmem,
this splits them into a separate record so that we can easily convert a
`Context.t` into a `Context.cacheable_t` and write that to sharedmem, then
convert a `Context.cacheable_t` back into a `Context.t` by re-applying
`Options.t`.

Reviewed By: avikchaudhuri

Differential Revision: D5153834

fbshipit-source-id: 787792ae18fb105a643d965e556943f5549cf498